requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.2
psycopg2-binary==2.9.6
beautifulsoup4==4.12.0
requests==2.31.0
python-dotenv==1.0.0